+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
         BmiService service = new BmiService();
-        double index = service.calculate(1.87, 98);
-        System.out.println((int) index);
+        int index = service.calculate(1.87, 98);
+        System.out.println(index);
     }
 }
